requirements.txt
pydub==0.25.1
pygame==2.6.1
python-vlc==3.0.21203
